Move to examples directory
